ReviewRequest Finished, added methods to filter by creator and reviewer in User class. Mail duplication check added.

+++ src/main/java/com/telerik/peer/mappers/UserMapper.java
@@ -34,7 +34,7 @@ public class UserMapper {
         return user;
     }
 
-    public User fromDto(UserDto userDto, int id) {
+    public User fromDto(UserDto userDto, long id) {
         User user = userService.getById(id);
         dtoToObject(userDto, user);
         return user;
